Initial Commit
The graphing utility used is gi-centre-utils, a great graphing library
for processing. Thanks to them:)
This currently is working with virtual data with much of the
architecture that will be needed when an arduino is connected.
The end goal of this project is to have a working frequency generator
and oscilloscope that can be used in conjunction with certain circuit
boards in education targeted at the New Zealand NCEA Level 3 Physics AC
curriculum.